requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
djangorestframework>=3.14.0
django-filter>=23.2 
django-cors-headers>=4.2.0
channels>=4.0.0
daphne>=4.0.0 
channels-redis>=4.1.0
drf-spectacular>=0.26.4
djangorestframework-simplejwt>=5.3.0
factory-boy>=3.3.0
hypothesis>=6.82.0